Move entity service registration to async_setup in ntfy integration
